% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfm-fit.R
\name{tabfm_fit}
\alias{tabfm_fit}
\title{Fit tabfm Tabular Foundation Model}
\usage{
tabfm_fit(X, y, problem_type = "classification", ...)
}
\arguments{
\item{X}{x training data}

\item{y}{y training data}

\item{problem_type}{either "classification" or "regression"}

\item{...}{for tabfm additional hyperparameters}
}
\value{
a tabfm fit
}
\description{
Fit tabfm Tabular Foundation Model
}
