% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfm-pred.R
\name{tabfm_predict}
\alias{tabfm_predict}
\title{tabfm Predictions}
\usage{
tabfm_predict(fit, data, type = "raw", ...)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{a dataset}

\item{type}{'raw' or 'proba'}

\item{...}{for passing parameters to the predict method}
}
\value{
predictions
}
\description{
tabfm Predictions
}
